+++ libs/dh/shared/data-access-msw/src/index.ts
@@ -14,5 +14,5 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export { setupServiceWorker } from './lib/mockedServiceWorker';
+export { setupServiceWorker } from './lib/setup-service-worker';
 export { handlers, onUnhandledRequest } from './lib/handlers';
